reverting back agentservice related changes.
